Separate ViaVersion methods to avoid ClassNotFoundExceptions
